(add) Pass context of document model to annotation views
In order to save the annotations we need access to the document, because
that is where/how everything is being stored.  Therefore we pass it through
to the views as a paramater so we can save an added annotation.

router.js: Pass relevant document to paragraphsView

paragraphsView: Receive doc via options, pass along to paragraphView